Inverse Foreign Keys
This adds an optional `inverseOf` option on `hasMany` associations so that you can correctly map relationships in which the child model has a non-default association key.

I also expanded test scenario coverage of all the tests that use the HasManyHelper so that they also include every permutation of default and non-default keys for the hasMany association and its inverse.