New internal db interface that doesn't protect against mutations
Making a copy of the db.model for every operation is very expensive,
but is done so end users can safely get a reference and mutate it locally
without mutating the db.

However, internally, Mirage can get a huge performance win by copying the records
for every db operation. This comes at the cost of not being safe against accidental
mutations, but that's an acceptable consequence since it only impacts private APIs.