drop deprecated forms of create, createList and normalizedRequestAttrs (#1525)
BREAKING CHANGE: There are several places in Mirage's APIs that were
intended to be used with singularized versions of model names, but just
so happened to work if a non-singularized version was passed in.

This behavior was discovered during a refactor, and the non-singularized
versions were maintained to avoid breaking apps. Now that we're moving to
1.0, we're removing this deprecated/unintentional behavior.

- `server.create` and `server.createList` were coded to take a singularized
model name, e.g. `server.create('user')`. It just so happens that
`server.create('users')` also works. That pluralized version is now removed
from Mirage.

    If you're running the latest 0.x version you should see a deprecation
    message letting you know where to change it. Otherwise, it should be
    a pretty mechanic change from things like `server.create('users')`
    to `serer.create('user')`.

    Note this also applies to `server.createList` – the correct form
    is `server.createList('user', 3)`, and the pluralized form
    `server.createList('users', 3)` is now unsupported.

- `this.normalizedRequestAttrs` in a route handler optionally takes a
  modelName as an argument. This is if your URLs are non-standard and
  Mirage cannot guess the modelName from the URL path.

    In this case, you can call `this.normalizedRequestAttrs('blog-post')`
    to tell Mirage to expect the payload to be for a `blog-post` model.

    This API was intended to be used with dasherized names, because that's
    how compound model names are specified throughout Mirage when they
    are represented as strings.

    It just so happened that `this.normalizedRequestAttrs('blogPost')`
    also worked, by chance, until a refactor. So, that behavior was kept
    but now is being removed.

    The correct usage is `this.normalizedRequestAttrs('blog-post')`.
    Using the camelized version of the model name is no longer supported.

If either of these changes cause a ton of refactoring pain, we can try
to marshal some resources to help write a codemod. Please open an
issue if that's the case!