Upgrade and lock-down prettier version (#988)
This is an upgrade of prettier from 2.2.1 to 2.4.1.  I also locked down the package.json, as prettier recommends, because any change to the version can cause formatting differences so it's best to be intentional about upgrades.

This PR also includes the results of running the new version of prettier, which made a few formatting changes, but not a ton.