Update jest to 27.2.4 (#991)
This makes a major-version update of jest from 26.6.3 to 27.2.4.

The only breaking change that seems to have had an impact to us is that `jsdom` is no longer the default `testEnvironment`, it has now been changed to `node`.  For that reason, I needed to explicitly specify `testEnvironment: "jsdom"` for the internal tests.  The other tests had their environments specified already.