Update transitive dependencies (#996)
Now that all of our direct dependencies are updated, this PR updates all of the transitive dependencies (sub-dependencies).  I did this by deleting the `yarn.lock` file, and re-creating it.  This allows semver to do its thing and de-duplicate as many dependencies as possible, which actually reduces the total number of packages installed in the mirage repo from 1024 to 960.

This isn't something I'd recommend doing all the time, but every once in a while I think it's good cleanup to do.