Add more useful typescript type for `modelInstance.attrs` (#1002)
I've found that I sometimes need to get the attributes of a model, but typescript was typing as a generic unknown object.  I _think_ we can type it the same as the properties that are added directly to the model itself, based on the wording of https://miragejs.com/api/classes/model/#attrs.