Route handler return types (#1009)
This updates the types for `RouteHandler` to allow for any valid JSON response. Adds `number | string | boolean | null` as well as arrays of `object | number | string | boolean | null`.

This is necessary because endpoint can return more than just an object. For instance the project I'm currently working on adding mirage too has an endpoint that returns just a number. Which is totally valid however using the current types without this change causes typescript to emit an error.