Type afterCreate in factories
When creating a factory when using typescript, the afterCreate hook
could not be defined as it is not a key of the model. This commit adds
typing for adding the afterCreate hook.

Unfortunately the exact type of the model is not inferred in this
commit. In order to make the factory fully infer the model, we would
need to create a dependency between the factory type and the model type,
which is something that seems to have been avoided purposefully.

IMHO the any model instance works good enough for now.