Breaking: Make createServer async to support MSW. (#1104)
This is to support MSW startup, which is async.  It has the downside of also making startup using pretender also async, even though it doesn't need to be.  Perhaps there's a bit of an inversion we can do here, so that the interceptors import from miragejs core, and users interact only directly with the interceptors.  I'll try experimenting around with something like that, but in the meantime, we could release this as an alpha version in case anyone wants to experiment with mirage-msw.  To that end, this PR is targeting a new `next` branch, to keep the main branch clean in case we need to release something (unlikely, I know).