Avoid stringifying valid Response body types (#1110)
This only works with mirage-msw, because pretender _always_ tries to deal with the response body as a string or `ArrayBuffer`.  But, together with one small change in mirage-msw (PR incoming soon), we can allow streaming mock responses when using MSW.  

I needed to use `jest-fixed-jsdom`, because otherwise jest doesn't know what to do with `ReadableStream` even though it's been native to web and node since at least Node 18.  

I also tried to write an explicit test for ReadableStream using mirage-msw, but it's not possible right now, because MSW throws an error if you try to use the browser worker in node.js.  I may explore switching to vitest, to use its new browser mode, so we can run node tests in node, and browser tests in a real browser.  But that will need to come later.