Fix Response body TypeScript types (#1111)
Ugh, this is what we get for writing the types separately. `KeyType` is
not what I thought it was. We just want the set of all valid key types,
which can be strings or numbers or symbols. The fix I've made here is
basically how TypeScript itself does it, so it should be correct now.